Update index.js
A set of methods that extend the Buffer object to write various data types (such as signed and unsigned 8-bit, 16-bit, 32-bit integers, 64-bit big integers, and single-precision and double-precision floating-point numbers) in the buffer in different byte orders (LE for little endian and BE for big endian). These methods allow developers to manipulate memory data in binary format, thereby handling data storage and transmission at a low level. The code also contains some auxiliary functions (such as checkInt and checkIEEE754) to ensure the legality of write operations and prevent out-of-bounds or values ​​outside the allowed range.